% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota_share.R
\name{quota_share}
\alias{quota_share}
\title{Quota Share Treaty}
\usage{
quota_share(ceded, ceding_commission)
}
\arguments{
\item{ceded}{Ceded proportion.}

\item{ceding_commission}{Ceding commission rate.}
}
\description{
Quota Share Treaty
}
