% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treaties.R
\name{treaty_apply_claims}
\alias{treaty_apply_claims}
\title{Apply Treaty to Claims}
\usage{
treaty_apply_claims(treaty, claims)
}
\arguments{
\item{treaty}{A `treaty` object.}

\item{claims}{A data frame with a `loss` column.}
}
\description{
Apply Treaty to Claims
}
