% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treaties.R
\name{treaty_apply_premiums}
\alias{treaty_apply_premiums}
\title{Apply Treaty to Premiums}
\usage{
treaty_apply_premiums(treaty, premiums)
}
\arguments{
\item{treaty}{A `treaty` object.}

\item{premiums}{A data frame with a `premium` column.}
}
\description{
Apply Treaty to Premiums
}
